Added reading from stdin and writing to stdout
Now the message can be omitted, in which case the input will be read
from stdin. So one could pipe together encrypt and decrypt (if they
wanted, for whatever reason). Also, if writing to a terminal, now a
newline is appended to the end.

One of the side benefits of this is that I removed allocations entirely.